+++ docs-src/components/index.js
@@ -1,26 +1,27 @@
 export { Button } from './button';
-export { PushButton } from './push-button';
 export { ButtonGroup } from './button-group';
+export { PushButton } from './push-button';
 
 export * from './input';
 
 export { InfoBar } from './info-bar';
-export { NotificationCenter } from './notification-center';
 export { MessageBox } from './message-box';
+export { NotificationCenter } from './notification-center';
 export { Tooltip } from './tooltip';
 
 export { Dialog } from './dialog';
 export { Drawer } from './drawer';
+export { Grid } from './grid';
 export { Panel } from './panel';
 export { Popover } from './popover';
-export { Grid } from './grid';
 export { Table } from './table';
 export { Tree } from './tree';
 
-export { Menu } from './menu';
+export { ColorPalette } from './color-palette';
 export { Icon } from './icon';
+export { Menu } from './menu';
+export { Shortcut } from './shortcut';
+export { Splitter } from './splitter';
 export { Tag } from './tag';
 export { ThemeSwitcher } from './theme-switcher';
 export { Utils } from './utils';
-export { Splitter } from './splitter';
-export { ColorPalette } from './color-palette';